Prep for splitting bm_fullstack
